kuntur_antivacunas_modelo_audio/requirements.txt
Flask
python-dotenv
b2sdk
opencv-python
imageio
faster-whisper
torch
sentence-transformers
pymongo
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
fastembed
keybert
nltk
requests
gunicorn
firebase-admin